Notification page now shows username and time

+++ app/src/main/java/com/dev/watchrant/adapters/NotifItem.java
@@ -1,14 +1,14 @@
 package com.dev.watchrant.adapters;
 
 public class NotifItem {
-    private int created_time;
+    private long created_time;
     private String type; // comment_vote, comment_content, comment_mention, comment_discuss, content_vote, rant_sub
     private int read; // 1 read 0 unread
     private int rant_id;
     private int uid; // ID of the user that "fired" the notif
     private String username;
 
-    public NotifItem(int created_time, String type, int read, int rant_id, int uid, String username) {
+    public NotifItem(long created_time, String type, int read, int rant_id, int uid, String username) {
         this.created_time = created_time;
         this.type = type;
         this.read = read;
@@ -19,7 +19,7 @@ public class NotifItem {
     public String getUsername() {
         return username;
     }
-    public int getCreated_time() {
+    public long getCreated_time() {
         return created_time;
     }
